requirements.txt
python-dotenv==1.0.1
requests==2.32.3
langchain>=0.2.15,<0.3.0
langchain-community>=0.2.14,<0.3.0
langchain-text-splitters>=0.2.2,<0.3.0
faiss-cpu==1.12.0
langchain-text-splitters==0.2.2
